Updated stats.c with new netlist and fixed previous format error.
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@3330 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ vpr/SRC/timing/path_delay2.c
@@ -255,13 +255,13 @@ float print_critical_path_node(FILE * fp, t_linked_int * critical_path_node) {
 		inet =
 				block[iblk].pb->rr_graph[pb_graph_pin->pin_count_in_cluster].net_num;
 		inet = vpack_to_clb_net_mapping[inet];
-		fprintf(fp, "External-to-Block Net: #%d (%s).  Pins on net: %u.\n",
-			inet, g_clbs_nlist.net[inet].name, g_clbs_nlist.net[inet].nodes.size());
+		fprintf(fp, "External-to-Block Net: #%d (%s).  Pins on net: %d.\n",
+			inet, g_clbs_nlist.net[inet].name, (int) g_clbs_nlist.net[inet].nodes.size());
 	} else if (pb_graph_pin != NULL) {
 		inet =
 				block[iblk].pb->rr_graph[pb_graph_pin->pin_count_in_cluster].net_num;
-		fprintf(fp, "Internal Net: #%d (%s).  Pins on net: %u.\n", inet,
-			g_atoms_nlist.net[inet].name, g_atoms_nlist.net[inet].nodes.size());
+		fprintf(fp, "Internal Net: #%d (%s).  Pins on net: %d.\n", inet,
+			g_atoms_nlist.net[inet].name, (int) g_atoms_nlist.net[inet].nodes.size());
 	}
 
 	fprintf(fp, "\n");
